requirements.txt
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.11.0
bcrypt==4.3.0
certifi==2025.4.26
click==8.1.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.7
orjson==3.10.15
pydantic==2.10.6
pydantic-extra-types==2.10.4
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.5
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.44.0
typer==0.15.3
typing_extensions==4.13.2
tzlocal==5.2
ujson==5.10.0
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
